Rename services names
