+++ camunda-formio-tasklist-vue/src/models/SearchOptions.ts
@@ -1,0 +1,10 @@
+export interface SearchOptionPayload {
+    label: string;
+    compares: Array<string>;
+    key: string;
+    values: Array<string>;
+    type: string;
+    variable?: string;
+    name: string;
+    operator: string;
+}
