+++ camunda-formio-tasklist-vue/src/models/index.ts
@@ -5,6 +5,7 @@ export * from "./FormRequestActionPayload";
 export * from "./FormRequestPayload";
 export * from "./GroupListPayload";
 export * from "./Payload";
+export * from "./SearchOptions";
 export * from "./SortOptions";
 export * from "./TaskHistoryListPayload";
 export * from "./TaskPayload";
